Dashboard & Compact Calendar
Changed up the dashboard and made it look more appealing for students and admins. Added a new compact calendar which is shows on the dashboard. Login from database works now. Fixed up the hamburger menu and made it more logical. Removed navbar completely as it was repetitive, and unncessary adding clutter.