Merge of Student and Users database + createuser changes.
Student and users tables in our database have been merged. Login now works with students instead of users which means alll previous users have been removed. The only current user working is John "mypassword" and Dora "mypassword". John is an admin, Dora is a student. There is one problem. for create users when asked for studentId for validation there is a post command getting a 405 error.